+++ Public/Get/Backup/Get-NinjaOneLocationBackupUsage.ps1
@@ -22,10 +22,10 @@ function Get-NinjaOneLocationBackupUsage {
     Param(
         # Organisation ID
         [Parameter(ValueFromPipelineByPropertyName, Mandatory, Position = 0)]
-        [Alias('id')]
         [Int]$organisationId,
         # Location ID
         [Parameter(ValueFromPipelineByPropertyName, Position = 1)]
+        [Alias('id')]
         [Int]$locationId
     )
     try {