+++ Public/Set/Set-NinjaOneDeviceCustomFields.ps1
@@ -9,18 +9,19 @@ function Set-NinjaOneDeviceCustomFields {
         .EXAMPLE
             PS> Set-NinjaOneDeviceCustomFields -deviceId 1 -customFields @{ CustomField1 = 'Value1'; CustomField2 = 'Value2' }
             
-            Set `CustomField1` to `Value1` and `CustomField2` to `Value2` respectively.
+            Set `CustomField1` to `Value1` and `CustomField2` to `Value2` respectively for the device with id 1.
     #>
-    [CmdletBinding( SupportsShouldProcess = $true, ConfirmImpact = 'Medium' )]
+    [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'Medium')]
     [OutputType([Object])]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSReviewUnusedParameter', '', Justification = 'Uses dynamic parameter parsing.')]
     Param(
         # The device to set the custom field value(s) for.
-        [Parameter(Mandatory = $true)]
+        [Parameter(Mandatory)]
         [string[]]$deviceId,
         # The custom field(s) body object.
-        [Parameter(Mandatory = $true)]
-        [object]$customFields
+        [Parameter(Mandatory)]
+        [Alias('customFields', 'body')]
+        [object]$deviceCustomFields
     )
     try {
         $Resource = "v2/device/$deviceId/custom-fields"